fix: get the first few predicted values instead of last few predicted values
Previously, if the sequence length was shortened, while the code was
correctly retrieving the last few values for the input sequence, it was erroneously
retrieving the last few values for the output sequence. In other words, given
the predicted output sequence is [1,2,3,4,5], and we only wanted to predict 3 values,
we were predicting [3,4,5] instead of [1,2,3].

This commit fixes this issue.